[Bronze II] Title: 숫자, Time: 1384 ms, Memory: 45208 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -1,24 +1,23 @@
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Scanner;
 
 // 숫자
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt(), b = sc.nextInt();
+        long a = sc.nextLong(), b = sc.nextLong();
         if (a > b) {
-            int tmp = a;
+            long tmp = a;
             a = b;
             b = tmp;
         }
-        ArrayList<Integer> list = new ArrayList();
-        for (int i = a + 1; i < b; i++) {
+        ArrayList<Long> list = new ArrayList();
+        for (long i = a + 1; i < b; i++) {
             list.add(i);
         }
         System.out.println(list.size());
 
-        for (Integer i : list) {
+        for (Long i : list) {
             System.out.print(i + " ");
         }
     }
